@@ -10,11 +10,10 @@ namespace App\Database;
 
 
 use App\Database\Resolve\ResolutionInterface;
-use Symfony\Component\Console\Command\Command;
+use Exception;
+use ReflectionClass;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\Console\Question\ChoiceQuestion;
-use Symfony\Component\Console\Question\ConfirmationQuestion;
 use Symfony\Component\Console\Style\SymfonyStyle;
 
 class DatabaseChecks
@@ -22,24 +21,30 @@ class DatabaseChecks
     /**
      * @var DatabaseCheck[]
      */
-    private $checks;
+    private iterable $checks;
 
-    public function __construct()
+    public function __construct(iterable $databaseChecks)
     {
-        $this->checks = [];
+        $this->checks = $databaseChecks;
     }
 
-    public function addCheck(DatabaseCheck $check)
-    {
-        $this->checks[] = $check;
-    }
-
-    public function runChecks(Command $command, InputInterface $input, OutputInterface $output)
+    public function runChecks(InputInterface $input, OutputInterface $output)
     {
         $io = new SymfonyStyle($input, $output);
 
+        $checks = iterator_to_array($this->checks);
+
+        // filter by limit
+        $limit = $input->getOption('limit');
+        if ($limit) {
+            $checks = array_filter($checks, function($check) use ($limit) {
+                $className = (new ReflectionClass($check))->getShortName();
+                return $className === $limit;
+            });
+        }
+
         // sort checks by priority, highest will be executed first
-        usort($this->checks, function(DatabaseCheck $a, DatabaseCheck $b) {
+        usort($checks, function(DatabaseCheck $a, DatabaseCheck $b) {
             if ($a->getPriority() == $b->getPriority()) {
                 return 0;
             }
@@ -47,55 +52,18 @@ class DatabaseChecks
             return ($a->getPriority() > $b->getPriority()) ? -1 : 1;
         });
 
-        foreach ($this->checks as $check) {
+        foreach ($checks as $check) {
             try {
-                $class = new \ReflectionClass($check);
+                $class = new ReflectionClass($check);
                 $io->section('Running check: ' . $class->getShortName());
 
-                $problems = $check->findProblems($io, $input->getOption('limit'));
-
-                // check is ok
-                if (sizeof($problems) === 0) {
-                    $io->success('Check finished without any problems.');
-                    continue;
-                }
-
-                // check found problems
-                $io->warning('Check found ' . sizeof($problems) . ' problems!');
-
-                // prepare choice question for resolution
-                $helper = $command->getHelper('question');
-
-                $resolutions = ['skip'];
-                $resolutionStrategies = $check->getResolutionStrategies();
-                foreach ($resolutionStrategies as $resolutionStrategy) {
-                    $resolutions[] = $resolutionStrategy->getKey();
-                }
-
-                $resolutionQuestion = new ChoiceQuestion('How do you want to resolve the issues?', $resolutions, 0);
-                $resolutionQuestion->setErrorMessage('Resolution %s is invalid.');
-
-                $pickedResolution = $helper->ask($input, $output, $resolutionQuestion);
-
-                // handle skipping
-                if ($pickedResolution === 'skip') {
-                    $io->text('skipping...');
-                    continue;
-                }
-
-                // resolve with strategy
-                $io->text('trying to resolve...');
-
-                $strategies = array_filter($resolutionStrategies, function(ResolutionInterface $strategy) use ($pickedResolution) {
-                    return $strategy->getKey() === $pickedResolution;
-                });
-                if ($strategies[0]->resolve($problems)) {
-                    $io->success('Check resolved problems');
+                if ($check->resolve($io)) {
+                    $io->success('Check resolved problems or nothing to fix');
                 } else {
                     $io->warning('Check failed resolving problems, aborting...');
                     break;
                 }
-            } catch (\Exception $e) {
+            } catch (Exception $e) {
                 $io->error($e->getMessage());
             }
 
