@@ -6,7 +6,9 @@ use ApiPlatform\Core\Annotation\ApiResource;
 use App\Controller\Api\GetAccountsCheckLocalLogin;
 use App\Controller\Api\GetAccountsWorkspaces;
 use App\Dto\LocalLoginInput;
+use DateTime;
 use Doctrine\ORM\Mapping as ORM;
+use InvalidArgumentException;
 use Symfony\Bridge\Doctrine\Validator\Constraints\UniqueEntity;
 use Symfony\Component\Security\Core\Encoder\EncoderAwareInterface;
 use Symfony\Component\Security\Core\User\UserInterface;
@@ -91,6 +93,12 @@ use Symfony\Component\Validator\Context\ExecutionContextInterface;
  */
 class Account implements UserInterface, EncoderAwareInterface, \Serializable
 {
+    public const ACTIVITY_ACTIVE = 'active';
+    public const ACTIVITY_ACTIVE_NOTIFIED = 'active_notified';
+    public const ACTIVITY_IDLE = 'idle';
+    public const ACTIVITY_IDLE_NOTIFIED = 'idle_notified';
+    public const ACTIVITY_ABANDONED = 'abandoned';
+
     /**
      * @var int|null
      *
@@ -111,7 +119,7 @@ class Account implements UserInterface, EncoderAwareInterface, \Serializable
      *
      * @Groups({"api_check_local_login"})
      */
-    private $contextId;
+    private int $contextId;
 
     /**
      * @var string
@@ -123,7 +131,7 @@ class Account implements UserInterface, EncoderAwareInterface, \Serializable
      *
      * @Groups({"api", "api_check_local_login"})
      */
-    private $username;
+    private string $username;
 
     /**
      * @Assert\NotBlank()
@@ -141,7 +149,7 @@ class Account implements UserInterface, EncoderAwareInterface, \Serializable
      *
      * @ORM\Column(type="string", length=32, nullable=true)
      */
-    private $passwordMd5;
+    private ?string $passwordMd5;
 
     /**
      * @ORM\Column(type="string", length=255, nullable=true)
@@ -150,54 +158,54 @@ class Account implements UserInterface, EncoderAwareInterface, \Serializable
      *
      * @Groups({"api_check_local_login"})
      */
-    private $password;
+    private ?string $password;
 
     /**
      * @var string
      *
-     * @ORM\Column(type="string", length=50, nullable=false)
+     * @ORM\Column(type="string", length=50)
      *
      * @Assert\NotBlank()
      *
      * @Groups({"api"})
      */
-    private $firstname;
+    private string $firstname;
 
     /**
      * @var string
      *
-     * @ORM\Column(type="string", length=50, nullable=false)
+     * @ORM\Column(type="string", length=50)
      *
      * @Assert\NotBlank()
      *
      * @Groups({"api"})
      */
-    private $lastname;
+    private string $lastname;
 
     /**
      * @var string
      *
-     * @ORM\Column(name="email", type="string", length=100, nullable=false)
+     * @ORM\Column(name="email", type="string", length=100)
      *
      * @Assert\Email()
      * @Assert\Callback({"App\Entity\Account", "validateMailRegex"})
      *
      * @Groups({"api"})
      */
-    private $email;
+    private string $email;
 
     /**
      * @var string
      *
-     * @ORM\Column(name="language", type="string", length=10, nullable=false)
+     * @ORM\Column(name="language", type="string", length=10)
      */
-    private $language;
+    private string $language;
 
     /**
      * @ORM\ManyToOne(targetEntity="App\Entity\AuthSource")
-     * @ORM\JoinColumn(nullable=false)
+     * @ORM\JoinColumn()
      */
-    private $authSource;
+    private AuthSource $authSource;
 
     /**
      * @var bool
@@ -206,7 +214,36 @@ class Account implements UserInterface, EncoderAwareInterface, \Serializable
      *
      * @Groups({"api"})
      */
-    private $locked = false;
+    private bool $locked = false;
+
+    /**
+     * @var DateTime|null
+     *
+     * @ORM\Column(name="last_login", type="datetime", nullable=true)
+     */
+    private ?DateTime $lastLogin;
+
+    /**
+     * @var string
+     *
+     * @ORM\Column(name="activity_state", type="string", length=15, options={"default"="active"})
+     */
+    private string $activityState;
+
+    /**
+     * @var DateTime|null
+     *
+     * @ORM\Column(name="activity_state_updated", type="datetime", nullable=true)
+     */
+    private ?DateTime $activityStateUpdated;
+
+    public function __construct()
+    {
+        $this->lastLogin = null;
+        $this->password = null;
+        $this->passwordMd5 = null;
+        $this->activityState = self::ACTIVITY_ACTIVE;
+    }
 
     /**
      * @return int|null
@@ -240,7 +277,7 @@ class Account implements UserInterface, EncoderAwareInterface, \Serializable
      *
      * @return (Role|string)[] The user roles
      */
-    public function getRoles()
+    public function getRoles(): array
     {
         return ['ROLE_USER'];
     }
@@ -253,7 +290,7 @@ class Account implements UserInterface, EncoderAwareInterface, \Serializable
      *
      * @return string The password
      */
-    public function getPassword()
+    public function getPassword(): string
     {
         return $this->password ?: $this->passwordMd5;
     }
@@ -275,7 +312,7 @@ class Account implements UserInterface, EncoderAwareInterface, \Serializable
      *
      * @return string|null The salt
      */
-    public function getSalt()
+    public function getSalt(): ?string
     {
         return null;
     }
@@ -303,7 +340,7 @@ class Account implements UserInterface, EncoderAwareInterface, \Serializable
      *
      * @return string The username
      */
-    public function getUsername()
+    public function getUsername(): string
     {
         return $this->username;
     }
@@ -436,12 +473,12 @@ class Account implements UserInterface, EncoderAwareInterface, \Serializable
     {
     }
 
-    public function hasLegacyPassword()
+    public function hasLegacyPassword(): bool
     {
         return $this->passwordMd5 !== null;
     }
 
-    public function getEncoderName()
+    public function getEncoderName(): ?string
     {
         if ($this->hasLegacyPassword()) {
             return 'legacy_encoder';
@@ -455,7 +492,7 @@ class Account implements UserInterface, EncoderAwareInterface, \Serializable
         return $this->authSource;
     }
 
-    public function setAuthSource(?AuthSource $authSource): self
+    public function setAuthSource(?AuthSource $authSource): Account
     {
         $this->authSource = $authSource;
 
@@ -474,14 +511,77 @@ class Account implements UserInterface, EncoderAwareInterface, \Serializable
      * @param bool $locked
      * @return Account
      */
-    public function setLocked(bool $locked): self
+    public function setLocked(bool $locked): Account
     {
         $this->locked = $locked;
         return $this;
     }
 
-    // Serializable
+    /**
+     * @return DateTime|null
+     */
+    public function getLastLogin(): ?DateTime
+    {
+        return $this->lastLogin;
+    }
 
+    /**
+     * @param DateTime|null $lastLogin
+     * @return Account
+     */
+    public function setLastLogin(?DateTime $lastLogin): Account
+    {
+        $this->lastLogin = $lastLogin;
+        return $this;
+    }
+
+    /**
+     * @return string
+     */
+    public function getActivityState(): string
+    {
+        return $this->activityState;
+    }
+
+    /**
+     * @param string $activityState
+     * @return Account
+     */
+    public function setActivityState(string $activityState): Account
+    {
+        if (!in_array($activityState, [
+            self::ACTIVITY_ACTIVE,
+            self::ACTIVITY_ACTIVE_NOTIFIED,
+            self::ACTIVITY_IDLE,
+            self::ACTIVITY_IDLE_NOTIFIED,
+            self::ACTIVITY_ABANDONED,
+        ])) {
+            throw new InvalidArgumentException("Invalid activity");
+        }
+
+        $this->activityState = $activityState;
+        return $this;
+    }
+
+    /**
+     * @return DateTime|null
+     */
+    public function getActivityStateUpdated(): ?DateTime
+    {
+        return $this->activityStateUpdated;
+    }
+
+    /**
+     * @param DateTime|null $activityStateUpdated
+     * @return Room
+     */
+    public function setActivityStateUpdated(?DateTime $activityStateUpdated): Account
+    {
+        $this->activityStateUpdated = $activityStateUpdated;
+        return $this;
+    }
+
+    // Serializable
     public function serialize()
     {
         $serializableData = get_object_vars($this);
