@@ -3,27 +3,24 @@
 namespace App\Entity;
 
 use ApiPlatform\Core\Annotation\ApiResource;
+use DateTime;
 use Doctrine\ORM\Mapping as ORM;
+use InvalidArgumentException;
 use OpenApi\Annotations as OA;
 use Symfony\Component\Serializer\Annotation\Groups;
-
 
 /**
  * Room
  *
  * @ORM\Table(name="room", indexes={
- *     @ORM\Index(name="context_id", columns={"context_id"}),
- *     @ORM\Index(name="creator_id", columns={"creator_id"}),
- *     @ORM\Index(name="type", columns={"type"}),
  *     @ORM\Index(name="activity", columns={"activity"}),
- *     @ORM\Index(name="deleter_id", columns={"deleter_id"}),
- *     @ORM\Index(name="deletion_date", columns={"deletion_date"}),
- *     @ORM\Index(name="room_description", columns={"room_description"}),
- *     @ORM\Index(name="contact_persons", columns={"contact_persons"}),
- *     @ORM\Index(name="title", columns={"title"}),
- *     @ORM\Index(name="modifier_id", columns={"modifier_id"}),
+ *     @ORM\Index(name="context_id", columns={"context_id"}),
+ *     @ORM\Index(name="delete_idx", columns={"deleter_id", "deletion_date"}),
+ *     @ORM\Index(name="lastlogin", columns={"lastlogin"}),
+ *     @ORM\Index(name="search_idx", columns={"title", "contact_persons"}),
  *     @ORM\Index(name="status", columns={"status"}),
- *     @ORM\Index(name="lastlogin", columns={"lastlogin"})
+ *     @ORM\Index(name="title", columns={"title"}),
+ *     @ORM\Index(name="type", columns={"type"})
  * })
  * @ORM\Entity(repositoryClass="App\Repository\RoomRepository")
  * @ORM\HasLifecycleCallbacks
@@ -45,6 +42,12 @@ use Symfony\Component\Serializer\Annotation\Groups;
  */
 class Room
 {
+    public const ACTIVITY_ACTIVE = 'active';
+    public const ACTIVITY_ACTIVE_NOTIFIED = 'active_notified';
+    public const ACTIVITY_IDLE = 'idle';
+    public const ACTIVITY_IDLE_NOTIFIED = 'idle_notified';
+    public const ACTIVITY_ABANDONED = 'abandoned';
+
     /**
      * @var integer
      *
@@ -55,158 +58,179 @@ class Room
      * @Groups({"api"})
      * @OA\Property(description="The unique identifier.")
      */
-    private $itemId = '0';
+    private $itemId;
 
     /**
-     * @var integer
+     * @var integer|null
      *
      * @ORM\Column(name="context_id", type="integer", nullable=true)
      */
-    private $contextId;
+    private ?int $contextId;
 
     /**
-     * @var integer
+     * @var User|null
      *
      * @ORM\ManyToOne(targetEntity="User")
-     * @ORM\JoinColumn(name="creator_id", referencedColumnName="item_id")
+     * @ORM\JoinColumn(name="creator_id", referencedColumnName="item_id", nullable=true)
      */
-    private $creator;
+    private ?User $creator;
 
     /**
-     * @var integer
+     * @var User|null
      *
      * @ORM\ManyToOne(targetEntity="User")
-     * @ORM\JoinColumn(name="modifier_id", referencedColumnName="item_id")
+     * @ORM\JoinColumn(name="modifier_id", referencedColumnName="item_id", nullable=true)
      */
-    private $modifier;
+    private ?User $modifier;
 
     /**
-     * @var integer
+     * @var User|null
      *
      * @ORM\ManyToOne(targetEntity="User")
      * @ORM\JoinColumn(name="deleter_id", referencedColumnName="item_id", nullable=true)
      */
-    private $deleter;
+    private ?User $deleter;
 
     /**
-     * @var \DateTime
+     * @var DateTime
      *
      * @ORM\Column(name="creation_date", type="datetime", nullable=false)
      *
      * @Groups({"api"})
      */
-    private $creationDate = '0000-00-00 00:00:00';
+    private DateTime $creationDate;
 
     /**
-     * @var \DateTime
+     * @var DateTime
      *
      * @ORM\Column(name="modification_date", type="datetime", nullable=false)
      *
      * @Groups({"api"})
      */
-    private $modificationDate = '0000-00-00 00:00:00';
+    private DateTime $modificationDate;
 
     /**
-     * @var \DateTime
+     * @var DateTime|null
      *
      * @ORM\Column(name="deletion_date", type="datetime", nullable=true)
      */
-    private $deletionDate;
+    private ?DateTime $deletionDate;
 
     /**
      * @var string
      *
-     * @ORM\Column(name="title", type="string", length=255, nullable=false)
+     * @ORM\Column(name="title", type="string", length=255)
      *
      * @Groups({"api"})
      * @OA\Property(type="string", maxLength=255)
      */
-    private $title;
+    private string $title;
 
     /**
-     * @var string
+     * @var array|null
      *
      * @ORM\Column(name="extras", type="mbarray", nullable=true)
      */
-    private $extras;
+    private ?array $extras = null;
 
     /**
      * @var string
      *
-     * @ORM\Column(name="status", type="string", length=20, nullable=false)
+     * @ORM\Column(name="status", type="string", length=20)
      */
-    private $status;
+    private string $status;
 
     /**
      * @var integer
      *
-     * @ORM\Column(name="activity", type="integer", nullable=false)
+     * @ORM\Column(name="activity", type="integer", options={"default":0})
      */
-    private $activity = '0';
+    private int $activity = 0;
 
     /**
      * @var string
      *
-     * @ORM\Column(name="type", type="string", length=20, nullable=false)
+     * @ORM\Column(name="type", type="string", length=20)
      *
      * @Groups({"api"})
      * @OA\Property(description="Either project or community")
      */
-    private $type = 'project';
+    private string $type = 'project';
 
     /**
      * @var boolean
      *
-     * @ORM\Column(name="public", type="boolean", nullable=false)
+     * @ORM\Column(name="public", type="boolean", options={"default":0})
      */
-    private $public = '0';
+    private bool $public = false;
 
     /**
      * @var boolean
      *
-     * @ORM\Column(name="is_open_for_guests", type="boolean", nullable=false)
+     * @ORM\Column(name="is_open_for_guests", type="boolean", options={"default":0})
      */
-    private $openForGuests = '0';
+    private bool $openForGuests = false;
 
     /**
-     * @var boolean
+     * @var int
      *
-     * @ORM\Column(name="continuous", type="boolean", nullable=false)
+     * @ORM\Column(name="continuous", type="smallint", options={"default":-1})
      */
-    private $continuous = '-1';
+    private int $continuous = -1;
 
     /**
-     * @var boolean
+     * @var int
      *
-     * @ORM\Column(name="template", type="integer", nullable=false)
+     * @ORM\Column(name="template", type="smallint", options={"default":-1})
      */
-    private $template = '-1';
+    private int $template = -1;
 
     /**
-     * @var string
+     * @var string|null
      *
      * @ORM\Column(name="contact_persons", type="string", length=255, nullable=true)
      */
-    private $contactPersons;
+    private ?string $contactPersons;
 
     /**
-     * @var string
+     * @var string|null
      *
      * @ORM\Column(name="room_description", type="string", length=10000, nullable=true)
      *
      * @Groups({"api"})
      * @OA\Property(type="string", nullable=true)
      */
-    private $roomDescription;
+    private ?string $roomDescription = null;
 
     /**
-     * @var \DateTime
+     * @var DateTime|null
      *
      * @ORM\Column(name="lastlogin", type="datetime", nullable=true)
      */
-    private $lastlogin;
+    private ?DateTime $lastlogin;
 
-    public function isIndexable()
+    /**
+     * @var string
+     *
+     * @ORM\Column(name="activity_state", type="string", length=15, options={"default"="active"})
+     */
+    private string $activityState;
+
+    /**
+     * @var DateTime|null
+     *
+     * @ORM\Column(name="activity_state_updated", type="datetime", nullable=true)
+     */
+    private ?DateTime $activityStateUpdated;
+
+    public function __construct()
+    {
+        $this->activityState = self::ACTIVITY_ACTIVE;
+        $this->creationDate = new DateTime();
+        $this->modificationDate = new DateTime();
+    }
+
+    public function isIndexable(): bool
     {
         return ($this->deleter == null && $this->deletionDate == null);
     }
@@ -222,7 +246,7 @@ class Room
         return 'user';
     }
 
-    public function setLanguage($language)
+    public function setLanguage($language): Room
     {
         $extras = $this->getExtras();
         $extras['LANGUAGE'] = $language;
@@ -231,12 +255,12 @@ class Room
         return $this;
     }
 
-    public function getLogo()
+    public function getLogo(): string
     {
         return '';
     }
 
-    public function getAccessCheck()
+    public function getAccessCheck(): string
     {
         $extras = $this->getExtras();
 
@@ -257,7 +281,7 @@ class Room
         return 'always';
     }
 
-    public function setAccessCheck($access)
+    public function setAccessCheck($access): Room
     {
         $mapping = array(
             'never' => -1,
@@ -272,17 +296,17 @@ class Room
         return $this;
     }
 
-    public function isProjectRoom()
+    public function isProjectRoom(): bool
     {
         return $this->type === 'project';
     }
 
-    public function isCommunityRoom()
+    public function isCommunityRoom(): bool
     {
         return $this->type === 'community';
     }
 
-    public function isMaterialOpenForGuests()
+    public function isMaterialOpenForGuests(): bool
     {
         $extras = $this->getExtras();
         if (isset($extras['MATERIAL_GUESTS'])) {
@@ -294,7 +318,7 @@ class Room
         return false;
     }
 
-    public function setIsMaterialOpenForGuests($open)
+    public function setIsMaterialOpenForGuests($open): Room
     {
         $extras = $this->getExtras();
         $extras['MATERIAL_GUESTS'] = $open;
@@ -303,7 +327,7 @@ class Room
         return $this;
     }
 
-    public function isAssignmentRestricted()
+    public function isAssignmentRestricted(): bool
     {
         $extras = $this->getExtras();
         if (isset($extras['ROOMASSOCIATION'])) {
@@ -315,7 +339,7 @@ class Room
         return false;
     }
 
-    public function setAssignmentRestricted($isRestricted)
+    public function setAssignmentRestricted($isRestricted): Room
     {
         $roomAssociation = 'forall';
 
@@ -326,6 +350,8 @@ class Room
         $extras = $this->getExtras();
         $extras['ROOMASSOCIATION'] = $roomAssociation;
         $this->setExtras($extras);
+
+        return $this;
     }
 
     /**
@@ -333,7 +359,7 @@ class Room
      *
      * @return integer
      */
-    public function getItemId()
+    public function getItemId(): int
     {
         return $this->itemId;
     }
@@ -345,7 +371,7 @@ class Room
      *
      * @return Room
      */
-    public function setContextId($contextId)
+    public function setContextId($contextId): Room
     {
         $this->contextId = $contextId;
 
@@ -357,7 +383,7 @@ class Room
      *
      * @return integer
      */
-    public function getContextId()
+    public function getContextId(): ?int
     {
         return $this->contextId;
     }
@@ -374,11 +400,11 @@ class Room
     /**
      * Set creationDate
      *
-     * @param \DateTime $creationDate
+     * @param DateTime $creationDate
      *
      * @return Room
      */
-    public function setCreationDate($creationDate)
+    public function setCreationDate($creationDate): Room
     {
         $this->creationDate = $creationDate;
 
@@ -388,9 +414,9 @@ class Room
     /**
      * Get creationDate
      *
-     * @return \DateTime
+     * @return DateTime
      */
-    public function getCreationDate()
+    public function getCreationDate(): DateTime
     {
         return $this->creationDate;
     }
@@ -398,11 +424,11 @@ class Room
     /**
      * Set modificationDate
      *
-     * @param \DateTime $modificationDate
+     * @param DateTime $modificationDate
      *
      * @return Room
      */
-    public function setModificationDate($modificationDate)
+    public function setModificationDate($modificationDate): Room
     {
         $this->modificationDate = $modificationDate;
 
@@ -412,9 +438,9 @@ class Room
     /**
      * Get modificationDate
      *
-     * @return \DateTime
+     * @return DateTime
      */
-    public function getModificationDate()
+    public function getModificationDate(): DateTime
     {
         return $this->modificationDate;
     }
@@ -422,11 +448,11 @@ class Room
     /**
      * Set deletionDate
      *
-     * @param \DateTime $deletionDate
+     * @param DateTime $deletionDate
      *
      * @return Room
      */
-    public function setDeletionDate($deletionDate)
+    public function setDeletionDate(DateTime $deletionDate): Room
     {
         $this->deletionDate = $deletionDate;
 
@@ -436,9 +462,9 @@ class Room
     /**
      * Get deletionDate
      *
-     * @return \DateTime
+     * @return DateTime
      */
-    public function getDeletionDate()
+    public function getDeletionDate(): ?DateTime
     {
         return $this->deletionDate;
     }
@@ -450,7 +476,7 @@ class Room
      *
      * @return Room
      */
-    public function setTitle($title)
+    public function setTitle(string $title): Room
     {
         $this->title = $title;
 
@@ -462,7 +488,7 @@ class Room
      *
      * @return string
      */
-    public function getTitle()
+    public function getTitle(): string
     {
         return $this->title;
     }
@@ -470,11 +496,11 @@ class Room
     /**
      * Set extras
      *
-     * @param mbarray $extras
+     * @param array $extras
      *
      * @return Room
      */
-    public function setExtras($extras)
+    public function setExtras(array $extras): Room
     {
         $this->extras = $extras;
 
@@ -484,9 +510,9 @@ class Room
     /**
      * Get extras
      *
-     * @return mbarray
+     * @return array|null
      */
-    public function getExtras()
+    public function getExtras(): ?array
     {
         return $this->extras;
     }
@@ -498,7 +524,7 @@ class Room
      *
      * @return Room
      */
-    public function setStatus($status)
+    public function setStatus($status): Room
     {
         $this->status = $status;
 
@@ -510,7 +536,7 @@ class Room
      *
      * @return string
      */
-    public function getStatus()
+    public function getStatus(): string
     {
         return $this->status;
     }
@@ -522,7 +548,7 @@ class Room
      *
      * @return Room
      */
-    public function setActivity($activity)
+    public function setActivity($activity): Room
     {
         $this->activity = $activity;
 
@@ -534,7 +560,7 @@ class Room
      *
      * @return integer
      */
-    public function getActivity()
+    public function getActivity(): int
     {
         return $this->activity;
     }
@@ -546,7 +572,7 @@ class Room
      *
      * @return Room
      */
-    public function setType($type)
+    public function setType($type): Room
     {
         $this->type = $type;
 
@@ -558,7 +584,7 @@ class Room
      *
      * @return string
      */
-    public function getType()
+    public function getType(): string
     {
         return $this->type;
     }
@@ -570,7 +596,7 @@ class Room
      *
      * @return Room
      */
-    public function setPublic($public)
+    public function setPublic($public): Room
     {
         $this->public = $public;
 
@@ -582,7 +608,7 @@ class Room
      *
      * @return boolean
      */
-    public function getPublic()
+    public function getPublic(): bool
     {
         return $this->public;
     }
@@ -594,7 +620,7 @@ class Room
      *
      * @return Room
      */
-    public function setOpenForGuests($openForGuests)
+    public function setOpenForGuests($openForGuests): Room
     {
         $this->openForGuests = $openForGuests;
 
@@ -606,7 +632,7 @@ class Room
      *
      * @return boolean
      */
-    public function getOpenForGuests()
+    public function getOpenForGuests(): bool
     {
         return $this->openForGuests;
     }
@@ -614,11 +640,11 @@ class Room
     /**
      * Set continuous
      *
-     * @param boolean $continuous
+     * @param int $continuous
      *
      * @return Room
      */
-    public function setContinuous($continuous)
+    public function setContinuous($continuous): Room
     {
         $this->continuous = $continuous;
 
@@ -628,9 +654,9 @@ class Room
     /**
      * Get continuous
      *
-     * @return boolean
+     * @return int
      */
-    public function getContinuous()
+    public function getContinuous(): int
     {
         return $this->continuous;
     }
@@ -638,13 +664,13 @@ class Room
     /**
      * Set template
      *
-     * @param boolean $template
+     * @param bool $template
      *
      * @return Room
      */
-    public function setTemplate($template)
+    public function setTemplate(bool $template): Room
     {
-        $this->template = $template;
+        $this->template = $template ? 1 : -1;
 
         return $this;
     }
@@ -652,11 +678,11 @@ class Room
     /**
      * Get template
      *
-     * @return boolean
+     * @return bool
      */
-    public function getTemplate()
+    public function getTemplate(): bool
     {
-        return $this->template;
+        return $this->template == 1;
     }
 
     /**
@@ -666,7 +692,7 @@ class Room
      *
      * @return Room
      */
-    public function setContactPersons($contactPersons)
+    public function setContactPersons($contactPersons): Room
     {
         $this->contactPersons = $contactPersons;
 
@@ -678,7 +704,7 @@ class Room
      *
      * @return string
      */
-    public function getContactPersons()
+    public function getContactPersons(): ?string
     {
         return $this->contactPersons;
     }
@@ -690,7 +716,7 @@ class Room
      *
      * @return Room
      */
-    public function setRoomDescription($roomDescription)
+    public function setRoomDescription($roomDescription): Room
     {
         $this->roomDescription = $roomDescription;
 
@@ -702,7 +728,7 @@ class Room
      *
      * @return string
      */
-    public function getRoomDescription()
+    public function getRoomDescription(): ?string
     {
         return $this->roomDescription;
     }
@@ -710,11 +736,11 @@ class Room
     /**
      * Set lastlogin
      *
-     * @param \DateTime $lastlogin
+     * @param DateTime $lastlogin
      *
      * @return Room
      */
-    public function setLastlogin($lastlogin)
+    public function setLastlogin($lastlogin): Room
     {
         $this->lastlogin = $lastlogin;
 
@@ -724,9 +750,9 @@ class Room
     /**
      * Get lastlogin
      *
-     * @return \DateTime
+     * @return DateTime
      */
-    public function getLastlogin()
+    public function getLastlogin(): ?DateTime
     {
         return $this->lastlogin;
     }
@@ -734,11 +760,11 @@ class Room
     /**
      * Set creator
      *
-     * @param \App\Entity\User $creator
+     * @param User|null $creator
      *
      * @return Room
      */
-    public function setCreator(\App\Entity\User $creator = null)
+    public function setCreator(User $creator = null): Room
     {
         $this->creator = $creator;
 
@@ -748,9 +774,9 @@ class Room
     /**
      * Get creator
      *
-     * @return \App\Entity\User
+     * @return User
      */
-    public function getCreator()
+    public function getCreator(): ?User
     {
         return $this->creator;
     }
@@ -758,11 +784,11 @@ class Room
     /**
      * Set modifier
      *
-     * @param \App\Entity\User $modifier
+     * @param User|null $modifier
      *
      * @return Room
      */
-    public function setModifier(\App\Entity\User $modifier = null)
+    public function setModifier(User $modifier = null): Room
     {
         $this->modifier = $modifier;
 
@@ -772,9 +798,9 @@ class Room
     /**
      * Get modifier
      *
-     * @return \App\Entity\User
+     * @return User
      */
-    public function getModifier()
+    public function getModifier(): ?User
     {
         return $this->modifier;
     }
@@ -782,11 +808,11 @@ class Room
     /**
      * Set deleter
      *
-     * @param \App\Entity\User $deleter
+     * @param User|null $deleter
      *
      * @return Room
      */
-    public function setDeleter(\App\Entity\User $deleter = null)
+    public function setDeleter(User $deleter = null): Room
     {
         $this->deleter = $deleter;
 
@@ -796,10 +822,56 @@ class Room
     /**
      * Get deleter
      *
-     * @return \App\Entity\User
+     * @return User|null
      */
-    public function getDeleter()
+    public function getDeleter(): ?User
     {
         return $this->deleter;
+    }
+
+    /**
+     * @return string
+     */
+    public function getActivityState(): string
+    {
+        return $this->activityState;
+    }
+
+    /**
+     * @param string $activityState
+     * @return Room
+     */
+    public function setActivityState(string $activityState): Room
+    {
+        if (!in_array($activityState, [
+            self::ACTIVITY_ACTIVE,
+            self::ACTIVITY_ACTIVE_NOTIFIED,
+            self::ACTIVITY_IDLE,
+            self::ACTIVITY_IDLE_NOTIFIED,
+            self::ACTIVITY_ABANDONED,
+        ], true)) {
+            throw new InvalidArgumentException("Invalid activity");
+        }
+
+        $this->activityState = $activityState;
+        return $this;
+    }
+
+    /**
+     * @return DateTime|null
+     */
+    public function getActivityStateUpdated(): ?DateTime
+    {
+        return $this->activityStateUpdated;
+    }
+
+    /**
+     * @param DateTime|null $activityStateUpdated
+     * @return Room
+     */
+    public function setActivityStateUpdated(?DateTime $activityStateUpdated): Room
+    {
+        $this->activityStateUpdated = $activityStateUpdated;
+        return $this;
     }
 }
