@@ -27,6 +27,7 @@ use App\Event\CommsyEditEvent;
 use App\Facade\UserCreatorFacade;
 use App\Form\Type\CsvImportType;
 use App\Form\Type\Portal\AccessibilityType;
+use App\Form\Type\Portal\AccountInactiveType;
 use App\Form\Type\Portal\AccountIndexDeleteUserType;
 use App\Form\Type\Portal\AccountIndexDetailAssignWorkspaceType;
 use App\Form\Type\Portal\AccountIndexDetailChangePasswordType;
@@ -38,7 +39,6 @@ use App\Form\Type\Portal\AccountIndexSendMailType;
 use App\Form\Type\Portal\AccountIndexSendMergeMailType;
 use App\Form\Type\Portal\AccountIndexSendPasswordMailType;
 use App\Form\Type\Portal\AccountIndexType;
-use App\Form\Type\Portal\ArchiveRoomsType;
 use App\Form\Type\Portal\AuthGuestType;
 use App\Form\Type\Portal\AuthLdapType;
 use App\Form\Type\Portal\AuthLocalType;
@@ -46,8 +46,8 @@ use App\Form\Type\Portal\AuthShibbolethType;
 use App\Form\Type\Portal\CommunityRoomsCreationType;
 use App\Form\Type\Portal\DataPrivacyType;
 use App\Form\Type\Portal\DeleteArchiveRoomsType;
+use App\Form\Type\Portal\GeneralType;
 use App\Form\Type\Portal\ImpressumType;
-use App\Form\Type\Portal\InactiveType;
 use App\Form\Type\Portal\LicenseSortType;
 use App\Form\Type\Portal\LicenseType;
 use App\Form\Type\Portal\MailtextsType;
@@ -59,6 +59,7 @@ use App\Form\Type\Portal\PortalhomeType;
 use App\Form\Type\Portal\PrivacyType;
 use App\Form\Type\Portal\ProjectRoomsCreationType;
 use App\Form\Type\Portal\RoomCategoriesType;
+use App\Form\Type\Portal\RoomInactiveType;
 use App\Form\Type\Portal\ServerAnnouncementsType;
 use App\Form\Type\Portal\ServerAppearanceType;
 use App\Form\Type\Portal\ServerGeneralType;
@@ -71,6 +72,7 @@ use App\Form\Type\TranslationType;
 use App\Mail\Mailer;
 use App\Model\TimePulseTemplate;
 use App\Repository\AuthSourceRepository;
+use App\Room\RoomManager;
 use App\Security\Authorization\Voter\RootVoter;
 use App\Services\LegacyEnvironment;
 use App\Services\RoomCategoriesService;
@@ -958,63 +960,53 @@ class PortalSettingsController extends AbstractController
      * @param Portal $portal
      * @param Request $request
      * @param EntityManagerInterface $entityManager
+     * @param AccountManager $accountManager
+     * @param RoomManager $roomManager
+     * @return array|RedirectResponse
      */
-    public function inactive(Portal $portal, Request $request, EntityManagerInterface $entityManager)
-    {
-        $inactiveForm = $this->createForm(InactiveType::class, $portal);
-
-        $inactiveForm->handleRequest($request);
-        if ($inactiveForm->isSubmitted() && $inactiveForm->isValid()) {
-
-            if ($inactiveForm->getClickedButton()->getName() === 'save') {
-                $entityManager->persist($portal);
-                $entityManager->flush();
-
-                return $this->redirectToRoute('app_portalsettings_inactive', [
-                    'portalId' => $portal->getId(),
-                    'tab' => 'inactive',
-                ]);
+    public function inactive(
+        Portal $portal,
+        Request $request,
+        EntityManagerInterface $entityManager,
+        AccountManager $accountManager,
+        RoomManager $roomManager
+    ) {
+        $accountInactiveForm = $this->createForm(AccountInactiveType::class, $portal);
+        $accountInactiveForm->handleRequest($request);
+        if ($accountInactiveForm->isSubmitted() && $accountInactiveForm->isValid()) {
+            // Reset all account if the feature has been disabled
+            if (!$portal->isClearInactiveAccountsFeatureEnabled()) {
+                $accountManager->resetInactivityToPreviousNonNotificationState();
             }
 
-            // TODO: inform the user how many inactive accounts would be locked/deleted due to the currently entered day values (see `configuration_inactive.php`)
+            $entityManager->persist($portal);
+            $entityManager->flush();
+
+            return $this->redirectToRoute('app_portalsettings_inactive', [
+                'portalId' => $portal->getId(),
+                'tab' => 'inactiveAccounts',
+            ]);
         }
 
-        // archiving rooms form
-        $archiveRoomsForm = $this->createForm(ArchiveRoomsType::class, $portal, []);
-        $archiveRoomsForm->handleRequest($request);
-        if ($archiveRoomsForm->isSubmitted() && $archiveRoomsForm->isValid()) {
-
-            if ($archiveRoomsForm->getClickedButton()->getName() === 'save') {
-                $entityManager->persist($portal);
-                $entityManager->flush();
-
-                return $this->redirectToRoute('app_portalsettings_inactive', [
-                    'portalId' => $portal->getId(),
-                    'tab' => 'archiveRooms',
-                ]);
+        $roomInactiveForm = $this->createForm(RoomInactiveType::class, $portal, []);
+        $roomInactiveForm->handleRequest($request);
+        if ($roomInactiveForm->isSubmitted() && $roomInactiveForm->isValid()) {
+            if (!$portal->isClearInactiveRoomsFeatureEnabled()) {
+                $roomManager->resetInactivityToPreviousNonNotificationState();
             }
-        }
 
-        // deleting archived rooms form
-        $deleteArchiveRoomsForm = $this->createForm(DeleteArchiveRoomsType::class, $portal, []);
-        $deleteArchiveRoomsForm->handleRequest($request);
-        if ($deleteArchiveRoomsForm->isSubmitted() && $deleteArchiveRoomsForm->isValid()) {
+            $entityManager->persist($portal);
+            $entityManager->flush();
 
-            if ($deleteArchiveRoomsForm->getClickedButton()->getName() === 'save') {
-                $entityManager->persist($portal);
-                $entityManager->flush();
-
-                return $this->redirectToRoute('app_portalsettings_inactive', [
-                    'portalId' => $portal->getId(),
-                    'tab' => 'deleteRooms',
-                ]);
-            }
+            return $this->redirectToRoute('app_portalsettings_inactive', [
+                'portalId' => $portal->getId(),
+                'tab' => 'inactiveRooms',
+            ]);
         }
 
         return [
-            'inactiveForm' => $inactiveForm->createView(),
-            'archiveRoomsForm' => $archiveRoomsForm->createView(),
-            'deleteArchiveRoomsForm' => $deleteArchiveRoomsForm->createView(),
+            'inactiveAccountsForm' => $accountInactiveForm->createView(),
+            'inactiveRoomsForm' => $roomInactiveForm->createView(),
             'tab' => $request->query->has('tab') ? $request->query->get('tab') : 'inactive',
         ];
     }
@@ -1486,10 +1478,14 @@ class PortalSettingsController extends AbstractController
         foreach ($portalUsers as $portalUser) {
             if (!in_array($portalUser->getUserID(),
                     $alreadyIncludedUserIDs) and $portalUser->getContextID() == $portalId) {
+
+                if ($portalUser->getUserID() != 'cschoenf2') continue;
+
                 $userList[] = $portalUser;
                 $alreadyIncludedUserIDs[] = $portalUser->getUserID();
             }
         }
+        unset($alreadyIncludedUserIDs);
 
         $accountIndex = new AccountIndex();
 
