@@ -47,16 +47,12 @@ class MailtextsType extends AbstractType
                     'Delete room' => 'MAIL_BODY_ROOM_DELETE',
                     'Create room' => 'MAIL_BODY_ROOM_OPEN',
                     '-------------------' => 28,
-                    'Archive room after X days' => 'PROJECT_MAIL_BODY_DELETE_INFO',
-                    'Room was archived yesterday' => 'PROJECT_MAIL_BODY_ARCHIVE',
-                    'Delete room after X day' => 'PROJECT_MAIL_BODY_DELETE_INFO',
-                    'Room was deleted yesterday' => 'PROJECT_MAIL_BODY_DELETE',
+                    'Lock room after X days' => 'EMAIL_INACTIVITY_ROOM_LOCK_UPCOMING_BODY',
+                    'Delete room after X day' => 'EMAIL_INACTIVITY_ROOM_DELETE_UPCOMING_BODY',
                     '--------------------' => 29,
                     'Lock userid in X days' => 'EMAIL_INACTIVITY_LOCK_NEXT_BODY',
-                    'Userid will be locked tomorrow' => 'EMAIL_INACTIVITY_LOCK_TOMORROW_BODY',
                     'Userid was locked' => 'EMAIL_INACTIVITY_LOCK_NOW_BODY',
                     'Delete userid in X days' => 'EMAIL_INACTIVITY_DELETE_NEXT_BODY',
-                    'Userid will be deleted tomorrow' => 'EMAIL_INACTIVITY_DELETE_TOMORROW_BODY',
                     'Userid was deleted' => 'EMAIL_INACTIVITY_DELETE_NOW_BODY',
                 ],
                 'required' => true,
@@ -71,6 +67,9 @@ class MailtextsType extends AbstractType
                 'label' => 'Content german',
                 'required' => false,
                 'translation_domain' => 'portal',
+                'attr' => [
+                    'rows' => 10,
+                ],
             ])
             ->add('resetContentGerman', CheckboxType::class, [
                 'label' => 'Reset',
@@ -80,6 +79,9 @@ class MailtextsType extends AbstractType
                 'label' => 'Content english',
                 'required' => false,
                 'translation_domain' => 'portal',
+                'attr' => [
+                    'rows' => 10,
+                ],
             ])
             ->add('resetContentEnglish', CheckboxType::class, [
                 'label' => 'Reset',
